banneredit >> bannercontroller
login 관련 == admin memberlogincontroller
홈페이지관리자 비밀번호변경, 등록 페이지 >> regist/adminmeberController

main.admdo
admin/main/adminMainController

main.do
user/main/UserMainController
