+++ src/main/webapp/WEB-INF/views/daminoAdminView/sites/banner/bannerBoardEdit.jsp
@@ -220,8 +220,8 @@
 								</table>
 								<br>
 								<div id="insert-menu-btns" align="center">
-									<input type="submit" class="btn btn-primary" value="메뉴등록" /> <input
-										type="button" class="btn btn-delete" value="목록" />
+									<input type="submit" class="btn btn-primary" value="메뉴등록" /> 
+									<input type="button" class="btn btn-delete" value="목록" onclick="location.href='bannerBoardView.admdo'" />
 								</div>
 							</div>
 						</form>
